+++ docs/demo/app/javascript/controllers/index.js
@@ -6,10 +6,11 @@ import { Application } from "@hotwired/stimulus"
 const application = Application.start()
 
 // Import LocoMotion controllers
-import { AlertController, CallyInputController, CountdownController, ThemeController } from "@profoundry-us/loco_motion"
+import { AlertController, CallyInputController, CountdownController, ModalController, ThemeController } from "@profoundry-us/loco_motion"
 application.register("loco-alert", AlertController)
 application.register("loco-cally-input", CallyInputController)
 application.register("loco-countdown", CountdownController)
+application.register("loco-modal", ModalController)
 application.register("loco-theme", ThemeController)
 
 // Import demo app controllers